[Bug fix] Don't add min_lovelace to unfulfilled_amount when change address is not provided.
Previously, in tx builder, we enforce input amount to be able to cover the minimum lovelace amount for change UTxO. This will cause a problem when user doesn't provide a change address. We fix this problem by checking whether a change address is provided before adding the minimum lovelace amount to unfulfilled amount.