Add warning when network type is passed to BlockFrostChainContext
The Cardano ledger specification only includes two network types: testnet and mainnet, which are utilized for address encoding. However, the BlockFrostChainContext should not rely on this network type as there are additional testnets, such as preview and preprod. To enhance the flexibility, the network parameter is being deprecated and replaced by directly passing the base_url for the network to the BlockFrostChainContext constructor.