Rename the reward redeemer tag to withdrawal
Ogmios 5 and blockfrost both refer to script redeemers that use the
withdrawal redeemer as "withdrawal:{index}". The only place where the
name of the redeemer tag is used is for evaluating the execution steps
in the transaction builder. Therefore I recommend renaming the redeemer
tag to the name used by other cardano tooling.

Note that in Ogmios 6, this name changes to "withdraw"