Fix fee too low
User can potentially provide signing keys that are not required by the finalized transaction.
For example, if a transaction include a minting script and signing key for the script, but the minted value is 0, the signature of the minting script isn't required.
This commit will auto detect such cases and skip the signing of this key.