Preserve container type of redeemers and plutus_data (#456)
Currently, fields in a witness will be forced to be wrapped in NonEmptyOrderedSet when deserialized from a cbor due to post_init.
This behavior will alter transaction id and also result in script_data_hash. This change is to ensure the original data format from a transaction will be preserved when deserialized.